Code migration: .NET Framework 4.8 to .NET 8.0 - convert csproj, Program.cs, appsettings.json, all controllers and views

+++ ContosoUniversity/Controllers/InstructorsController.cs
@@ -1,19 +1,19 @@
-using System;
-using System.Collections.Generic;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System.Linq;
-using System.Net;
-using System.Web.Mvc;
 using ContosoUniversity.Data;
 using ContosoUniversity.Models;
 using ContosoUniversity.Models.SchoolViewModels;
+using ContosoUniversity.Services;
 
 namespace ContosoUniversity.Controllers
 {
     public class InstructorsController : BaseController
     {
-        // GET: Instructors - All roles can view
-        public ActionResult Index(int? id, int? courseID)
+        public InstructorsController(SchoolContext context, NotificationService notificationService)
+            : base(context, notificationService) { }
+
+        // GET: Instructors
+        public IActionResult Index(int? id, int? courseID)
         {
             var viewModel = new InstructorIndexData();
             viewModel.Instructors = db.Instructors
@@ -40,23 +40,23 @@ namespace ContosoUniversity.Controllers
             return View(viewModel);
         }
 
-        // GET: Instructors/Details/5 - All roles can view details
-        public ActionResult Details(int? id)
+        // GET: Instructors/Details/5
+        public IActionResult Details(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors.Find(id);
+            Instructor? instructor = db.Instructors.Find(id);
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             return View(instructor);
         }
 
         // GET: Instructors/Create
-        public ActionResult Create()
+        public IActionResult Create()
         {
             var instructor = new Instructor();
             instructor.CourseAssignments = new List<CourseAssignment>();
@@ -67,7 +67,7 @@ namespace ContosoUniversity.Controllers
         // POST: Instructors/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "LastName,FirstMidName,HireDate,OfficeAssignment")] Instructor instructor, string[] selectedCourses)
+        public IActionResult Create([Bind("LastName,FirstMidName,HireDate,OfficeAssignment")] Instructor instructor, string[]? selectedCourses)
         {
             if (selectedCourses != null)
             {
@@ -78,38 +78,38 @@ namespace ContosoUniversity.Controllers
                     instructor.CourseAssignments.Add(courseToAdd);
                 }
             }
+
             if (ModelState.IsValid)
             {
                 db.Instructors.Add(instructor);
                 db.SaveChanges();
-                
-                // Send notification for instructor creation
+
                 SendEntityNotification("Instructor", instructor.ID.ToString(), EntityOperation.CREATE);
-                
-                return RedirectToAction("Index");
+
+                return RedirectToAction(nameof(Index));
             }
             PopulateAssignedCourseData(instructor);
             return View(instructor);
         }
 
         // GET: Instructors/Edit/5
-        public ActionResult Edit(int? id)
+        public IActionResult Edit(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors
+            Instructor? instructor = db.Instructors
                 .Include(i => i.OfficeAssignment)
                 .Include(i => i.CourseAssignments)
                     .ThenInclude(c => c.Course)
                 .Where(i => i.ID == id)
-                .Single();
-            PopulateAssignedCourseData(instructor);
+                .SingleOrDefault();
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
+            PopulateAssignedCourseData(instructor);
             return View(instructor);
         }
 
@@ -133,25 +133,30 @@ namespace ContosoUniversity.Controllers
         // POST: Instructors/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(int? id, string[] selectedCourses)
+        public IActionResult Edit(int? id, string[]? selectedCourses)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
             var instructorToUpdate = db.Instructors
                .Include(i => i.OfficeAssignment)
                .Include(i => i.CourseAssignments)
                    .ThenInclude(c => c.Course)
                .Where(i => i.ID == id)
-               .Single();
+               .SingleOrDefault();
 
-            if (TryUpdateModel(instructorToUpdate, "",
-               new string[] { "LastName", "FirstMidName", "HireDate", "OfficeAssignment" }))
+            if (instructorToUpdate == null)
+            {
+                return NotFound();
+            }
+
+            if (TryUpdateModelAsync(instructorToUpdate, "",
+               i => i.LastName, i => i.FirstMidName, i => i.HireDate, i => i.OfficeAssignment).Result)
             {
                 try
                 {
-                    if (String.IsNullOrWhiteSpace(instructorToUpdate.OfficeAssignment.Location))
+                    if (string.IsNullOrWhiteSpace(instructorToUpdate.OfficeAssignment?.Location))
                     {
                         instructorToUpdate.OfficeAssignment = null;
                     }
@@ -159,11 +164,10 @@ namespace ContosoUniversity.Controllers
                     UpdateInstructorCourses(selectedCourses, instructorToUpdate);
 
                     db.SaveChanges();
-                    
-                    // Send notification for instructor update
+
                     SendEntityNotification("Instructor", instructorToUpdate.ID.ToString(), EntityOperation.UPDATE);
 
-                    return RedirectToAction("Index");
+                    return RedirectToAction(nameof(Index));
                 }
                 catch (Exception)
                 {
@@ -174,7 +178,7 @@ namespace ContosoUniversity.Controllers
             return View(instructorToUpdate);
         }
 
-        private void UpdateInstructorCourses(string[] selectedCourses, Instructor instructorToUpdate)
+        private void UpdateInstructorCourses(string[]? selectedCourses, Instructor instructorToUpdate)
         {
             if (selectedCourses == null)
             {
@@ -183,8 +187,8 @@ namespace ContosoUniversity.Controllers
             }
 
             var selectedCoursesHS = new HashSet<string>(selectedCourses);
-            var instructorCourses = new HashSet<int>
-                (instructorToUpdate.CourseAssignments.Select(c => c.Course.CourseID));
+            var instructorCourses = new HashSet<int>(
+                instructorToUpdate.CourseAssignments.Select(c => c.Course.CourseID));
             foreach (var course in db.Courses)
             {
                 if (selectedCoursesHS.Contains(course.CourseID.ToString()))
@@ -196,40 +200,47 @@ namespace ContosoUniversity.Controllers
                 }
                 else
                 {
-
                     if (instructorCourses.Contains(course.CourseID))
                     {
-                        CourseAssignment courseToRemove = instructorToUpdate.CourseAssignments.SingleOrDefault(i => i.CourseID == course.CourseID);
-                        db.Entry(courseToRemove).State = EntityState.Deleted;
+                        CourseAssignment? courseToRemove = instructorToUpdate.CourseAssignments.SingleOrDefault(i => i.CourseID == course.CourseID);
+                        if (courseToRemove != null)
+                        {
+                            db.Entry(courseToRemove).State = EntityState.Deleted;
+                        }
                     }
                 }
             }
         }
 
         // GET: Instructors/Delete/5
-        public ActionResult Delete(int? id)
+        public IActionResult Delete(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors.Find(id);
+            Instructor? instructor = db.Instructors.Find(id);
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             return View(instructor);
         }
 
-        // POST: Instructors/Delete/5 - Only admins can delete instructors
+        // POST: Instructors/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public IActionResult DeleteConfirmed(int id)
         {
-            Instructor instructor = db.Instructors
+            Instructor? instructor = db.Instructors
               .Include(i => i.OfficeAssignment)
               .Where(i => i.ID == id)
-              .Single();
+              .SingleOrDefault();
+
+            if (instructor == null)
+            {
+                return NotFound();
+            }
 
             db.Instructors.Remove(instructor);
 
@@ -242,20 +253,10 @@ namespace ContosoUniversity.Controllers
             }
 
             db.SaveChanges();
-            
-            // Send notification for instructor deletion
-            SendEntityNotification("Instructor", id.ToString(), EntityOperation.DELETE);
-            
-            return RedirectToAction("Index");
-        }
 
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
+            SendEntityNotification("Instructor", id.ToString(), EntityOperation.DELETE);
+
+            return RedirectToAction(nameof(Index));
         }
     }
 }
